(#4835) - test in phantomjs again
I have high hopes that PhantomJS 2 is gonna work. As a bonus,
we no longer need the ES5 shims, because PhantomJS supports
ES5 fully. However, it does have the broken IndexedDB, so it
will test in WebSQL instead (I confirmed that it's doing this).

Basically I am proposing this because I want to test WebSQL, but
I don't want to test it in Chrome or Safari over Saucelabs, because
that's slow and flaky.